Updated some OS based directory path code in SVNRepoConfig. Now using the Path.Combine and Path.DirectorySeparatorChar instead of hand checking the path.
Added a new constructor to SVNRepository to support passing a SVNRepoConfig object as a constructor parameter. This was done to support the use case where a user sets up svnserve to use a global configuration file. The changes did not break the code, but the new constructor has not been tested.

Added properties to hold values related to the global config file to SVNServerConfig.cs.